feat: refactor mixin so that both v3 and v4 can use it
